Kerala_Support/requirements.txt
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.0.2
django-cors-headers==2.2.0
django-modeladmin-reorder==0.3.1
djangorestframework==3.7.7
idna==2.6
Pillow==5.1.0
psycopg2==2.7.4
pycparser==2.18
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
xlrd==1.1.0
